+++ central/src/auth/authentik/app.rs
@@ -6,9 +6,24 @@ use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use shared::{OIDCConfig, SecretResult};
 
-use super::{generate_secret, group, AuthentikConfig};
+use super::{generate_secret, get_access_token, group, AuthentikConfig};
 
-async fn generate_application(provider: &str, name: &str, oidc_client_config: &OIDCConfig, secret: &str, conf: &AuthentikConfig, token: &str)
+pub async fn create_application(
+    name: &str,
+    oidc_client_config: OIDCConfig,
+    conf: &AuthentikConfig,
+) -> reqwest::Result<SecretResult> {
+    let token = get_access_token(conf).await?;
+    post_application(&token, name, &oidc_client_config, conf).await
+}
+
+async fn generate_application(
+    provider: &str, 
+    name: &str, 
+    oidc_client_config: &OIDCConfig, 
+    secret: &str, 
+    conf: &AuthentikConfig, 
+    token: &str)
 -> Result<(), String>
 {
     let secret = (!oidc_client_config.is_public).then_some(secret);
@@ -80,7 +95,7 @@ async fn generate_provider(name: &str, oidc_client_config: &OIDCConfig, secret: 
     }
 }
 
-pub fn post_application(
+pub async fn post_application(
     token: &str,
     name: &str,
     oidc_client_config: &OIDCConfig,
@@ -91,51 +106,9 @@ pub fn post_application(
     } else {
         String::with_capacity(0)
     };
-    let generated_provider = generate_provider(name, oidc_client_config, &secret, conf, token);
-    let generated_app = generate_application(name, name, oidc_client_config, &secret, conf);
+    //let generated_provider = generate_provider(name, oidc_client_config, &secret, conf, token).await;
+    let generated_app = generate_application(name, name, oidc_client_config, &secret, conf, token).await;
     // Create groups for this client
-    create_groups(name, token, conf).await?;
-    match res.status() {
-        StatusCode::CREATED => {
-            println!("Client for {name} created.");
-            if !oidc_client_config.is_public {
-                let client_id = generated_app
-                    .get("clientId")
-                    .and_then(Value::as_str)
-                    .expect("Always present");
-            }
-            Ok(SecretResult::Created(secret))
-        }
-        StatusCode::CONFLICT => {
-            let conflicting_client = get_client(name, token, oidc_client_config, conf).await?;
-            if client_configs_match(&conflicting_client, &generated_client) {
-                Ok(SecretResult::AlreadyExisted(conflicting_client
-                    .as_object()
-                    .and_then(|o| o.get("secret"))
-                    .and_then(Value::as_str)
-                    .unwrap_or("")
-                    .to_owned()))
-            } else {
-                Ok(CLIENT
-                    .put(&format!(
-                        "{}/admin/realms/{}/clients/{}",
-                        conf.keycloak_url,
-                        conf.keycloak_realm,
-                        conflicting_client
-                            .get("clientId")
-                            .and_then(Value::as_str)
-                            .expect("We have a valid client")
-                    ))
-                    .bearer_auth(token)
-                    .json(&generated_client)
-                    .send()
-                    .await?
-                    .status()
-                    .is_success()
-                    .then_some(SecretResult::AlreadyExisted(secret))
-                    .expect("We know the client already exists so updating should be successful"))
-            }
-        }
-        s => unreachable!("Unexpected statuscode {s} while creating keycloak client. {res:?}"),
-    }
+    let generated_group = create_groups(name, token, conf).await?;
+
 }
